+++ packages/logger/src/logger.ts
@@ -4,6 +4,7 @@ import stringify from 'fast-safe-stringify';
 import * as symbols from './symbols';
 import sanitize from './sanitize';
 import getDurationString from './util/duration';
+import hrtimestamp from './util/timestamp';
 import ensureOptions, { LogOptions, LogLevel } from './options';
 
 // Nice clean log level definitions
@@ -48,7 +49,7 @@ export type JSONLog = {
   message: Array<string | object | any>;
   level: LogFns;
   name?: string;
-  time: number;
+  time: string;
 };
 
 export type StringLog = [LogFns | 'confirm' | 'print', ...any];
@@ -174,10 +175,9 @@ export default function (name?: string, options: LogOptions = {}): Logger {
       level,
       name,
       message,
-      time: Date.now(),
+      time: hrtimestamp().toString(),
     };
 
-    stringify(output);
     emitter[level](stringify(output));
   };