+++ packages/ws-worker/src/events.ts
@@ -36,7 +36,7 @@ export type ATTEMPT_COMPLETE_REPLY = undefined;
 export const ATTEMPT_LOG = 'attempt:log'; // level, namespace (job,runtime,adaptor), message, time
 export type ATTEMPT_LOG_PAYLOAD = {
   message: Array<string | object>;
-  timestamp: number;
+  timestamp: string; // Tiemstamp in microseconds (13 digits)
   attempt_id: string;
   level?: string;
   source?: string; // namespace