+++ packages/deploy/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/deploy
 
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies [ca701e8]
+  - @openfn/logger@0.0.19
+
 ## 0.2.8
 
 ### Patch Changes