+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @openfn/runtime
 
+## 0.0.33
+
+### Patch Changes
+
+- Removed log line
+- Updated dependencies [ca701e8]
+  - @openfn/logger@0.0.19
+
 ## 0.0.32
 
 ### Patch Changes